+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/db_list/DatabaseListCommandLineParser.java
@@ -60,6 +60,11 @@ public class DatabaseListCommandLineParser {
 		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.DB_LIST;
 
+		if (cmd.hasOption("verbose"))
+			result.verbosity = 2;
+		if (cmd.hasOption("very-verbose"))
+			result.verbosity = 3;
+
 		if (cmd.hasOption("help")) {
 			printHelp();
 			throw new HelpRequestedException();