+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **develop**
 
+* Adding `-v` and `-vv` command line options.
 * Using log4j/slf4j for I/O in jannovar-core.
 * Adding `PrintStream` as parameter to `JannovarOptions#print`.
 * Compressing serialized file.